@@ -1,70 +1,48 @@
 {
   "name": "8398a7.github.io",
-  "version": "1.0.0",
-  "main": "index.js",
-  "repository": "ssh://git@github.com/8398a7/8398a7.github.io.git",
-  "author": "839 <8398a7@gmail.com>",
-  "license": "MIT",
-  "scripts": {
-    "start": "webpack-dev-server --progress --colors --hot",
-    "docker:start": "docker build . -t 8398a7/8398a7.github.io:latest && docker run -it -v `pwd`/src:/usr/src/app/src -p 8080:8080 8398a7/8398a7.github.io:latest",
-    "staging:build": "rm -rf dist/* && cp development.html dist/index.html && NODE_ENV=staging webpack",
-    "build": "rm -rf dist/* && cp development.html dist/index.html && NODE_ENV=production webpack",
-    "lint": "tslint -c tslint.json 'src/**/*.ts*'",
-    "lint:fix": "tslint --fix -c tslint.json 'src/**/*.ts*'"
-  },
-  "devDependencies": {
-    "@types/history": "^4.6.0",
-    "@types/react": "^15.0.29",
-    "@types/react-dom": "^15.5.0",
-    "@types/react-hot-loader": "^3.0.1",
-    "@types/react-redux": "^5.0.1",
-    "@types/react-router": "^4.0.11",
-    "@types/react-router-dom": "^4.0.4",
-    "@types/react-router-redux": "^5.0.2",
-    "@types/redux": "^3.6.0",
-    "@types/redux-immutable": "^3.0.33",
-    "@types/redux-logger": "^3.0.0",
-    "@types/webpack": "^3.0.8",
-    "@types/webpack-env": "^1.13.0",
-    "awesome-typescript-loader": "^3.2.2",
-    "css-loader": "^0.28.4",
-    "extract-text-webpack-plugin": "^2.1.2",
-    "file-loader": "^0.11.2",
-    "node-env-file": "^0.1.8",
-    "node-sass": "^4.5.3",
-    "optimize-css-assets-webpack-plugin": "^2.0.0",
-    "sass-loader": "^6.0.6",
-    "style-loader": "^0.18.2",
-    "tslint": "^5.6.0",
-    "tslint-config-airbnb": "^5.2.1",
-    "tslint-react": "^3.2.0",
-    "typescript": "^2.3.4",
-    "uglifyjs-webpack-plugin": "^0.4.6",
-    "webpack": "^3.5.4",
-    "webpack-dev-server": "^2.7.1",
-    "webpack-sentry-plugin": "^1.9.2"
-  },
+  "version": "0.1.0",
+  "private": true,
   "dependencies": {
-    "axios": "^0.16.2",
+    "@sentry/browser": "^4.0.6",
+    "@types/raven-for-redux": "^1.1.1",
+    "@types/raven-js": "^3.10.0",
+    "@types/react-redux": "^6.0.9",
+    "@types/react-router-dom": "^4.3.1",
+    "@types/redux": "^3.6.0",
+    "@types/redux-immutable": "^4.0.0",
+    "@types/redux-logger": "^3.0.6",
+    "action-reducer": "^0.2.0",
+    "axios": "^0.18.0",
+    "connected-react-router": "^4.5.0",
     "font-awesome": "^4.7.0",
-    "history": "^4.6.3",
-    "immutable": "^3.8.1",
-    "materialize-css": "^0.98.2",
-    "react": "next",
-    "react-dom": "next",
-    "react-ga": "^2.2.0",
-    "react-hot-loader": "next",
-    "react-redux": "^5.0.6",
-    "react-router": "^4.1.1",
-    "react-router-dom": "^4.1.1",
-    "react-router-redux": "next",
-    "react-twitter-widgets": "^1.3.0",
-    "redux": "^3.7.2",
+    "immutable": "^3.8.2",
+    "materialize-css": "^1.0.0",
+    "raven-for-redux": "^1.3.1",
+    "react": "^16.5.2",
+    "react-dom": "^16.5.2",
+    "react-ga": "^2.5.3",
+    "react-redux": "^5.0.7",
+    "react-router-dom": "^4.3.1",
+    "react-scripts-ts": "3.1.0",
+    "react-twitter-widgets": "^1.7.1",
+    "redux": "^4.0.0",
     "redux-immutable": "^4.0.0",
     "redux-logger": "^3.0.6",
-    "redux-raven-middleware": "^1.2.0",
-    "redux-saga": "^0.15.6",
-    "video-react": "^0.8.6"
+    "redux-saga": "^0.16.2",
+    "video-react": "^0.13.0"
+  },
+  "scripts": {
+    "start": "react-scripts-ts start",
+    "build": "react-scripts-ts build",
+    "test": "react-scripts-ts test --env=jsdom",
+    "eject": "react-scripts-ts eject",
+    "publish": "react-scripts-ts build && cp -r ./build/* ."
+  },
+  "devDependencies": {
+    "@types/jest": "^23.3.4",
+    "@types/node": "^10.11.5",
+    "@types/react": "^16.4.16",
+    "@types/react-dom": "^16.0.9",
+    "typescript": "^3.1.1"
   }
 }
