@@ -1,8 +1,8 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
-import './Header.scss';
+import profile from '../images/profile.jpg';
+import './Header.css';
 import Navbar from './Navbar';
-const profile: string = require('../images/profile.jpg');
 
 const Header: React.StatelessComponent<{}> = () => (
   <header>
